moved clean function in standalone cs
