+++ AI Helper/AI Helper.csproj
@@ -190,6 +190,7 @@
     <Compile Include="Games\Illusion\HoneySelect.cs" />
     <Compile Include="Games\Illusion\AiGirl.cs" />
     <Compile Include="Games\GameBase.cs" />
+    <Compile Include="Manage\Functions\DataDirCleaner.cs" />
     <Compile Include="Manage\ManageHardlinks.cs" />
     <Compile Include="Manage\ManageReport.cs" />
     <Compile Include="Manage\ModeSwitch\GameBackuper.cs" />
