+++ AI Helper/Forms/MainForm.cs
@@ -13,6 +13,7 @@ using AIHelper.Install.Types;
 using AIHelper.Install.Types.Directories;
 using AIHelper.Install.UpdateMaker;
 using AIHelper.Manage;
+using AIHelper.Manage.Functions;
 using AIHelper.Manage.Update;
 using AIHelper.Manage.Update.Targets;
 using CheckForEmptyDir;
@@ -90,7 +91,7 @@ namespace AIHelper
                     return;
                 }
 
-                await new Updater().Update(new List<UpdateTargetBase>() {new Mo(new UpdateInfo()) }).ConfigureAwait(true);
+                await new Updater().Update(new List<UpdateTargetBase>() { new Mo(new UpdateInfo()) }).ConfigureAwait(true);
             }
 
             ManageSettings.MOIsNew = IsMo23OrNever();
@@ -1563,7 +1564,7 @@ namespace AIHelper
 
         private void CleanDataDirLabel_Click(object sender, EventArgs e)
         {
-            ManageOther.CleanCurrentGameDataDir();
+            new DataDirCleaner().Clean();
         }
 
         //Disable close window button
